# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "This is an API for ToGo List application",
        "title": "ToGo List API",
        "contact": {},
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8080",
    "basePath": "/v1",
    "paths": {
        "/healthz": {
            "get": {
                "description": "Health Checking for API services",
                "produces": [
                    "application/json"
                ],
                "summary": "Health Checking",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.HealthzResult"
                        }
                    }
                }
            }
        },
        "/lists": {
            "get": {
                "description": "Get all Lists",
                "produces": [
                    "application/json"
                ],
                "summary": "Get All Lists",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "type": "array",
                            "items": {
                                "$ref": "#/definitions/dto.List"
                            }
                        }
                    }
                }
            },
            "post": {
                "description": "Create a single List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Create List",
                "parameters": [
                    {
                        "description": "Create List",
                        "name": "list",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.CreateListInput"
                        }
                    }
                ],
                "responses": {
                    "201": {
                        "description": "Created",
                        "schema": {
                            "$ref": "#/definitions/dto.List"
                        }
                    }
                }
            }
        },
        "/lists/{id}": {
            "get": {
                "description": "Get a single List",
                "produces": [
                    "application/json"
                ],
                "summary": "Get One List",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "List ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.List"
                        }
                    },
                    "404": {
                        "description": "List not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "put": {
                "description": "Update a single List",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "summary": "Update List",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "List ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "Update List",
                        "name": "list",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/dto.UpdateListInput"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/dto.List"
                        }
                    },
                    "404": {
                        "description": "List not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            },
            "delete": {
                "description": "Delete a single List",
                "produces": [
                    "application/json"
                ],
                "summary": "Delete List",
                "parameters": [
                    {
                        "type": "integer",
                        "description": "List ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "List deleted",
                        "schema": {
                            "$ref": "#/definitions/dto.List"
                        }
                    },
                    "404": {
                        "description": "List not found",
                        "schema": {
                            "type": "string"
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "dto.CreateListInput": {
            "type": "object",
            "required": [
                "description",
                "title"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "This is my list"
                },
                "priority": {
                    "type": "integer",
                    "example": 0
                },
                "title": {
                    "type": "string",
                    "example": "My List"
                }
            }
        },
        "dto.HealthzResult": {
            "type": "object",
            "properties": {
                "details": {
                    "$ref": "#/definitions/dto.HealthzResultDetails"
                },
                "status": {
                    "type": "string",
                    "enum": [
                        "ok",
                        "error"
                    ],
                    "example": "ok"
                }
            }
        },
        "dto.HealthzResultDetail": {
            "type": "object",
            "properties": {
                "status": {
                    "type": "string",
                    "enum": [
                        "ok",
                        "error"
                    ],
                    "example": "ok"
                }
            }
        },
        "dto.HealthzResultDetails": {
            "type": "object",
            "properties": {
                "db": {
                    "$ref": "#/definitions/dto.HealthzResultDetail"
                }
            }
        },
        "dto.List": {
            "type": "object",
            "properties": {
                "description": {
                    "type": "string",
                    "example": "This is my list"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "priority": {
                    "type": "integer",
                    "example": 0
                },
                "title": {
                    "type": "string",
                    "example": "My List"
                }
            }
        },
        "dto.UpdateListInput": {
            "type": "object",
            "required": [
                "id"
            ],
            "properties": {
                "description": {
                    "type": "string",
                    "example": "This is my list"
                },
                "id": {
                    "type": "integer",
                    "example": 1
                },
                "priority": {
                    "type": "integer",
                    "example": 0
                },
                "title": {
                    "type": "string",
                    "example": "My List"
                }
            }
        }
    }
}